# documentation/final-project-api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/quizzes/{id}": {
            "get": {
                "tags": [
                    "quiz-controller"
                ],
                "operationId": "getQuizById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Quiz"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "quiz-controller"
                ],
                "operationId": "updateQuiz",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuizDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "quiz-controller"
                ],
                "operationId": "deleteQuiz",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/results/submit": {
            "post": {
                "tags": [
                    "quiz-result-controller"
                ],
                "operationId": "submitQuiz",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuizSubmissionDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/quizzes": {
            "get": {
                "tags": [
                    "quiz-controller"
                ],
                "operationId": "getAllQuizzes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/QuizSummaryDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "quiz-controller"
                ],
                "operationId": "createQuiz",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuizDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/results/{resultId}": {
            "get": {
                "tags": [
                    "quiz-result-controller"
                ],
                "operationId": "getResultById",
                "parameters": [
                    {
                        "name": "resultId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/QuizResult"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/results/user/{userId}": {
            "get": {
                "tags": [
                    "quiz-result-controller"
                ],
                "operationId": "getResultsForUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/QuizResult"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "QuizDTO": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 3
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "required": [
                    "title"
                ]
            },
...
            "QuizSubmissionDTO": {
                "type": "object",
                "properties": {
                    "quizId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "answers": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        },
                        "minProperties": 1
                    }
                },
                "required": [
                    "answers",
                    "quizId"
                ]
            },
...
            "Question": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "questionText": {
                        "type": "string",
                        "minLength": 1
                    },
                    "optionA": {
                        "type": "string",
                        "minLength": 1
                    },
                    "optionB": {
                        "type": "string",
                        "minLength": 1
                    },
                    "optionC": {
                        "type": "string"
                    },
                    "optionD": {
                        "type": "string"
                    },
                    "correctAnswer": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "correctAnswer",
                    "optionA",
                    "optionB",
                    "questionText"
                ]
            },
            "Quiz": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string",
                        "minLength": 1
                    },
                    "description": {
                        "type": "string"
                    },
                    "questions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Question"
                        }
                    }
                },
                "required": [
                    "title"
                ]
            },
            "QuizResult": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "quiz": {
                        "$ref": "#/components/schemas/Quiz"
                    },
                    "score": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "completedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "QuizSummaryDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "createdByUsername": {
                        "type": "string"
                    }
                }
            }
        }
    }
}